+++ lib/operator/curry.ts
@@ -1,3 +1,0 @@
-import curry from 'lodash/curry'
-
-export { curry }