+++ src/function/index.ts
@@ -1,3 +1,7 @@
 export function copy<T = any> (o: T) { 
   return Object.assign(Object.create(null), o) as T 
 }
+
+function clone <T = any> (o: T) {
+  
+}