+++ src/operator/curry.ts
@@ -1,0 +1,6 @@
+export function curry(fn: Function, ...args: any[]){
+  return (..._args: any[]) =>
+    (
+      rest => rest.length >= fn.length ? fn(...rest) : curry(fn, rest)
+    )([ ...args, ..._args ])
+}