+++ src/operator/operators.ts
@@ -1,7 +1,11 @@
 import { equal } from './equal'
 import { isObject } from './type'
+import { curry } from './curry'
 
-export const is = equal(undefined)
+export const is: <T = any>(a: T, b: T) => boolean = curry(equal, Object.is)
+export const not: <T = any>(a: T, b: T) => boolean = curry(equal, function () {
+  return !Object.is.apply(null, arguments)
+})
 
 export function deepEqual (a: any, b: any): boolean {
   if (a === b) return true
@@ -36,11 +40,9 @@ export function deepEqual (a: any, b: any): boolean {
   }
 }
 
-export const not = equal(function () {
-  return !Object.is.apply(null, arguments)
-})
 
 
-export const identity = (_: any) => _
+
+export const identity = <T = any>(value: T) => value
 export const noop = (..._args: any[]) => {}
 export const no = (..._args: any[]) => false